Menambah Tabel SQL Baru dan mengubah lokasi file model dan controller lama
