+++ app/Models/PastModels/Trip.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace App\Models;
+
+use App\Models\Toko;
+use App\Models\User;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+
+class Trip extends Model
+{
+    use HasFactory;
+
+    protected $guarded = ['id'];
+
+    public function linkToko()
+    {
+        return $this->belongsTo(Toko::class, 'id', 'toko_id');
+    }
+
+    public function linkSales()
+    {
+        return $this->hasOne(User::class, 'id', 'sales_id');
+    }
+}
